chore(condo) integrate canDelete to softDeleted.js
